LoongArch: Remove invalid 'd' suffix for double literals
Per C++ 14 specification, "2.14.4 Floating literals", 'd' is not a
standardized floating-suffix.
The 'd' floating-suffix is a GCC extension, which worked on GCC by
default but not supported by Clang and many other C compilers.

Having no suffix makes a constant of double, which is expected here, so
it can be removed safely.

Fixes: 09ae79f2af71 ("Update (2023.12.08)")
Signed-off-by: Bingwu Zhang <xtex@aosc.io>