LoongArch: Remove unused va_list to VMError::report_and_die
VMError::report_and_die provides an overload that takes a variable-length
argument as the last argument, so we can use it instead of the overload
taking va_list as the last argument.

The implementation detail of va_list is compiler-defined and
constructing a va_list manually is not standardized which may not work
as expected.

This should resolve the warning saying that report_and_die uses the
uninitialized variable of va_dummy.

Signed-off-by: Bingwu Zhang <xtex@aosc.io>