LoongArch: Avoid formatter injection
The sixth argument of VMError::report_and_die is a formatting string,
using detail_msg as the 6th arg leaves a location for formatter
injecting, although it is hard to be used because almost all detail_msg
point to static strings.

This should resolve the Clang warning saying that the formatter string
is not constant.

Signed-off-by: Bingwu Zhang <xtex@aosc.io>